UI Integration bug fixed + merged firmwareLink and register redux actions/reducers
