+++ src/components/App.js
@@ -64,7 +64,6 @@ class App extends React.PureComponent {
     } else if (this.props.isFetching && !this.props.isEthereumEnabled) {
       // begin enable ethereum process (default application state at beginning of user flow)
       // console.log('loading fetching')
-      {/*TODO make sure the loader is centered in the middle of the screen*/}
       appView = (<Loader loadingText={"Loading Código"}/>);
     } else {
       //  An error has occurred logging users in with ethereum
