+++ src/reducers/index.js
@@ -2,7 +2,6 @@ import { combineReducers } from 'redux';
 import auth from './auth';
 import alerts from './alerts';
 import ethereum from "./ethereum";
-import firmware from "./firmware";
 import navigation from './navigation';
 import register from './register';
 
@@ -10,7 +9,6 @@ export default combineReducers({
   alerts,
   auth,
   ethereum,
-  firmware,
   navigation,
   register
 });
